Move model files to commons: nip51Lists, nip01UserMetadata, nip17Dms, nip65RelayList, edits, localRelays, nip72Communities, nip78AppSpecific, nipA3PaymentTargets, serverList
Moved 28 model files from amethyst to commons/commonMain:
- nip51Lists: PinListState, BlockPeopleListState, BlockedRelayListState,
  BroadcastRelayListState, GeohashListState, HashtagListState,
  IndexerRelayListState, LabeledBookmarkList, MuteListState, PeopleList,
  ProxyRelayListState, RelayFeedListState, SearchRelayListState,
  TrustedRelayListState
- nip01UserMetadata: AccountHomeRelayState, AccountOutboxRelayState,
  NotificationInboxRelayState, UserMetadataState
- nip17Dms: DmInboxRelayState, DmRelayListState
- nip65RelayList: Nip65RelayListState (typealias to existing commons)
- edits: PrivateStorageRelayListState
- localRelays: ForwardKind0ToLocalRelayState, LocalRelayListState
- nip72Communities: CommunityListState
- nip78AppSpecific: AppSpecificState
- nipA3PaymentTargets: NipA3PaymentTargetsState
- serverList: TrustedRelayListsState

Key changes:
- LocalCache → ICacheProvider (constructor injection)
- Dispatchers.IO → Dispatchers_IO (KMP compatible)
- AccountSettings implements Nip65RelayListRepository
- Typealias files left at original locations for backward compatibility

Not moved (Android/service dependencies):
- nip02FollowLists (depends on topNavFeeds/OutboxRelayLoader)
- nip03Timestamp (OkHttp, java.util.Base64)
- nip51Lists/HiddenUsersState (service layer)
- nip51Lists/FollowListsState, PeopleListsState, LabeledBookmarkListsState (Account class)
- nip47WalletConnect/NwcSignerState (consume method)
- nip62Vanish/VanishRequestsState (observeNotes)
- nipB7Blossom/BlossomServerListState (UI imports)
- serverList/MergedFollow*RelayListsState (nip02 deps)
- trustedAssertions/TrustProviderListState (interface conflict)
