+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip01UserMetadata/AccountOutboxRelayState.kt
@@ -20,39 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip01UserMetadata
 
-import com.vitorpamplona.amethyst.model.edits.PrivateStorageRelayListState
-import com.vitorpamplona.amethyst.model.localRelays.LocalRelayListState
-import com.vitorpamplona.amethyst.model.nip51Lists.broadcastRelays.BroadcastRelayListState
-import com.vitorpamplona.amethyst.model.nip65RelayList.Nip65RelayListState
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.stateIn
-
-class AccountOutboxRelayState(
-    nip65: Nip65RelayListState,
-    privateStorage: PrivateStorageRelayListState,
-    local: LocalRelayListState,
-    broadcast: BroadcastRelayListState,
-    scope: CoroutineScope,
-) {
-    val flow =
-        combine(
-            nip65.outboxFlow,
-            privateStorage.flow,
-            local.flow,
-            broadcast.flow,
-        ) { nip65Outbox, privateOutBox, localRelays, broadcastRelays ->
-            nip65Outbox + privateOutBox + localRelays + broadcastRelays
-        }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                nip65.outboxFlow.value +
-                    privateStorage.flow.value +
-                    local.flow.value +
-                    broadcast.flow.value,
-            )
-}
+typealias AccountOutboxRelayState = com.vitorpamplona.amethyst.commons.model.nip01UserMetadata.AccountOutboxRelayState
