+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip01UserMetadata/UserMetadataState.kt
@@ -20,133 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip01UserMetadata
 
-import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.quartz.nip01Core.metadata.MetadataEvent
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip39ExtIdentities.ExternalIdentitiesEvent
-import com.vitorpamplona.quartz.utils.Log
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-
-class UserMetadataState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val scope: CoroutineScope,
-    val settings: AccountSettings,
-) {
-    // Creates a long-term reference for this note so that the GC doesn't collect the note it self
-    val user = cache.getOrCreateUser(signer.pubKey)
-
-    // Creates a long-term reference for this note so that the GC doesn't collect the note it self
-    val note = cache.getOrCreateAddressableNote(getMetadataEventAddress())
-
-    fun getMetadataEventAddress() = MetadataEvent.createAddress(signer.pubKey)
-
-    // fun getEphemeralChatListAddress() = cache.getOrCreateUser(signer.pubKey)
-
-    fun getUserMetadataFlow() = note.flow().metadata.stateFlow
-
-    fun getUserMetadataEvent(): MetadataEvent? = note.event as? MetadataEvent
-
-    fun getIdentitiesEventAddress() = ExternalIdentitiesEvent.createAddress(signer.pubKey)
-
-    val identitiesNote = cache.getOrCreateAddressableNote(getIdentitiesEventAddress())
-
-    fun getExternalIdentitiesEvent(): ExternalIdentitiesEvent? = identitiesNote.event as? ExternalIdentitiesEvent
-
-    suspend fun sendNewUserMetadata(
-        name: String? = null,
-        displayName: String? = null,
-        picture: String? = null,
-        banner: String? = null,
-        website: String? = null,
-        pronouns: String? = null,
-        about: String? = null,
-        nip05: String? = null,
-        lnAddress: String? = null,
-        lnURL: String? = null,
-    ): MetadataEvent {
-        val latest = getUserMetadataEvent()
-
-        val template =
-            if (latest != null) {
-                MetadataEvent.updateFromPast(
-                    latest = latest,
-                    name = name,
-                    displayName = displayName,
-                    picture = picture,
-                    banner = banner,
-                    website = website,
-                    pronouns = pronouns,
-                    about = about,
-                    nip05 = nip05,
-                    lnAddress = lnAddress,
-                    lnURL = lnURL,
-                )
-            } else {
-                MetadataEvent.createNew(
-                    name = name,
-                    displayName = displayName,
-                    picture = picture,
-                    banner = banner,
-                    website = website,
-                    pronouns = pronouns,
-                    about = about,
-                    nip05 = nip05,
-                    lnAddress = lnAddress,
-                    lnURL = lnURL,
-                )
-            }
-
-        return signer.sign(template)
-    }
-
-    suspend fun sendNewUserIdentities(
-        twitter: String? = null,
-        mastodon: String? = null,
-        github: String? = null,
-    ): ExternalIdentitiesEvent {
-        val latest = getExternalIdentitiesEvent()
-
-        val template =
-            if (latest != null) {
-                ExternalIdentitiesEvent.updateFromPast(
-                    latest = latest,
-                    twitter = twitter,
-                    mastodon = mastodon,
-                    github = github,
-                )
-            } else {
-                ExternalIdentitiesEvent.createNew(
-                    twitter = twitter,
-                    mastodon = mastodon,
-                    github = github,
-                )
-            }
-
-        return signer.sign(template)
-    }
-
-    init {
-        settings.backupUserMetadata?.let {
-            Log.d("AccountRegisterObservers") { "Loading saved user metadata ${it.toJson()}" }
-
-            @OptIn(DelicateCoroutinesApi::class)
-            scope.launch(Dispatchers.IO) { LocalCache.justConsumeMyOwnEvent(it) }
-        }
-
-        // saves contact list for the next time.
-        scope.launch(Dispatchers.IO) {
-            Log.d("AccountRegisterObservers", "Kind 0 Collector Start")
-            getUserMetadataFlow().collect {
-                Log.d("AccountRegisterObservers") { "Updating Kind 0 ${user.toBestDisplayName()}" }
-                (it.note.event as? MetadataEvent)?.let {
-                    settings.updateUserMetadata(it)
-                }
-            }
-        }
-    }
-}
+typealias UserMetadataState = com.vitorpamplona.amethyst.commons.model.nip01UserMetadata.UserMetadataState
