+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/PinListState.kt
@@ -20,111 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists
 
-import androidx.compose.runtime.Stable
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.NoteState
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.PinListEvent
-import com.vitorpamplona.quartz.nip51Lists.bookmarkList.tags.EventBookmark
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.FlowPreview
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.debounce
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-
-@Stable
-class PinListState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val scope: CoroutineScope,
-) {
-    val pinList = cache.getOrCreateAddressableNote(PinListEvent.createPinAddress(signer.pubKey))
-
-    fun getPinListFlow(): StateFlow<NoteState> = pinList.flow().metadata.stateFlow
-
-    fun getPinList(): PinListEvent? = pinList.event as? PinListEvent
-
-    fun pinnedEvents(note: Note): List<EventBookmark> {
-        val noteEvent = note.event as? PinListEvent
-        return noteEvent?.pinnedEvents() ?: emptyList()
-    }
-
-    @OptIn(FlowPreview::class)
-    val pinnedNotes: StateFlow<List<EventBookmark>> =
-        getPinListFlow()
-            .map { noteState ->
-                pinnedEvents(noteState.note)
-            }.onStart {
-                emit(pinnedEvents(pinList))
-            }.debounce(100)
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptyList(),
-            )
-
-    val pinnedEventIdSet: StateFlow<Set<String>> =
-        pinnedNotes
-            .map { pins ->
-                pins.map { it.eventId }.toSet()
-            }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-
-    @OptIn(FlowPreview::class)
-    val pinnedNotesList: StateFlow<List<Note>> =
-        pinnedNotes
-            .map { pins ->
-                pins.mapNotNull { cache.checkGetOrCreateNote(it.eventId) }.reversed()
-            }.onStart {
-                emit(
-                    pinnedNotes.value.mapNotNull { cache.checkGetOrCreateNote(it.eventId) }.reversed(),
-                )
-            }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptyList(),
-            )
-
-    fun isPinned(note: Note): Boolean = pinnedEventIdSet.value.contains(note.idHex)
-
-    suspend fun addPin(note: Note): PinListEvent {
-        val currentList = getPinList()
-        val pin = EventBookmark(note.idHex, note.relayHintUrl())
-
-        return if (currentList == null) {
-            PinListEvent.create(
-                pin = pin,
-                signer = signer,
-            )
-        } else {
-            PinListEvent.add(
-                earlierVersion = currentList,
-                pin = pin,
-                signer = signer,
-            )
-        }
-    }
-
-    suspend fun removePin(note: Note): PinListEvent? {
-        val currentList = getPinList() ?: return null
-        val pin = EventBookmark(note.idHex, note.relayHintUrl())
-
-        return PinListEvent.remove(
-            earlierVersion = currentList,
-            pin = pin,
-            signer = signer,
-        )
-    }
-}
+typealias PinListState = com.vitorpamplona.amethyst.commons.model.nip51Lists.PinListState
