+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/blockPeopleList/BlockPeopleListState.kt
@@ -20,101 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.blockPeopleList
 
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.NoteState
-import com.vitorpamplona.amethyst.model.nip51Lists.peopleList.PeopleListDecryptionCache
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.muteList.tags.MuteTag
-import com.vitorpamplona.quartz.nip51Lists.muteList.tags.UserTag
-import com.vitorpamplona.quartz.nip51Lists.muteList.tags.WordTag
-import com.vitorpamplona.quartz.nip51Lists.peopleList.PeopleListEvent
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-
-class BlockPeopleListState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val decryptionCache: PeopleListDecryptionCache,
-    val scope: CoroutineScope,
-) {
-    // Creates a long-term reference for this note so that the GC doesn't collect the note it self
-    val blockListNote = cache.getOrCreateAddressableNote(getBlockListAddress())
-
-    fun getBlockListAddress() = PeopleListEvent.createBlockAddress(signer.pubKey)
-
-    fun getBlockListFlow(): StateFlow<NoteState> = blockListNote.flow().metadata.stateFlow
-
-    fun getBlockList(): PeopleListEvent? = blockListNote.event as? PeopleListEvent
-
-    suspend fun blockListWithBackup(note: Note): List<MuteTag> {
-        val event = note.event as? PeopleListEvent
-        return event?.let { decryptionCache.usersAndWords(it) } ?: emptyList()
-    }
-
-    val flow =
-        getBlockListFlow()
-            .map { blockListWithBackup(it.note) }
-            .onStart { emit(blockListWithBackup(blockListNote)) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptyList(),
-            )
-
-    suspend fun hideUser(pubkeyHex: String): PeopleListEvent {
-        val blockList = getBlockList()
-
-        return if (blockList != null) {
-            PeopleListEvent.add(
-                earlierVersion = blockList,
-                person = UserTag(pubkeyHex),
-                isPrivate = true,
-                signer = signer,
-            )
-        } else {
-            PeopleListEvent.create(
-                title = PeopleListEvent.BLOCK_LIST_D_TAG,
-                person = UserTag(pubkeyHex),
-                isPrivate = true,
-                signer = signer,
-                dTag = PeopleListEvent.BLOCK_LIST_D_TAG,
-            )
-        }
-    }
-
-    suspend fun showUser(pubkeyHex: String): PeopleListEvent? {
-        val blockList = getBlockList()
-
-        return if (blockList != null) {
-            PeopleListEvent.remove(
-                earlierVersion = blockList,
-                person = UserTag(pubkeyHex),
-                signer = signer,
-            )
-        } else {
-            null
-        }
-    }
-
-    suspend fun showWord(word: String): PeopleListEvent? {
-        val blockList = getBlockList()
-
-        return if (blockList != null) {
-            PeopleListEvent.remove(
-                earlierVersion = blockList,
-                person = WordTag(word),
-                signer = signer,
-            )
-        } else {
-            null
-        }
-    }
-}
+typealias BlockPeopleListState = com.vitorpamplona.amethyst.commons.model.nip51Lists.blockPeopleList.BlockPeopleListState
