+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/blockedRelays/BlockedRelayListState.kt
@@ -20,92 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.blockedRelays
 
-import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.NoteState
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip01Core.signers.SignerExceptions
-import com.vitorpamplona.quartz.nip51Lists.relayLists.BlockedRelayListEvent
-import com.vitorpamplona.quartz.utils.Log
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.launch
-
-class BlockedRelayListState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val decryptionCache: BlockedRelayListDecryptionCache,
-    val scope: CoroutineScope,
-    val settings: AccountSettings,
-) {
-    // Creates a long-term reference for this note so that the GC doesn't collect the note it self
-    val blockedListNote = cache.getOrCreateAddressableNote(getBlockedRelayListAddress())
-
-    fun getBlockedRelayListAddress() = BlockedRelayListEvent.createAddress(signer.pubKey)
-
-    fun getBlockedRelayListFlow(): StateFlow<NoteState> = blockedListNote.flow().metadata.stateFlow
-
-    fun getBlockedRelayList(): BlockedRelayListEvent? = blockedListNote.event as? BlockedRelayListEvent
-
-    suspend fun normalizeBlockedRelayListWithBackup(note: Note): Set<NormalizedRelayUrl> {
-        val event = note.event as? BlockedRelayListEvent ?: settings.backupBlockedRelayList
-        return event?.let { decryptionCache.relays(it) } ?: emptySet()
-    }
-
-    val flow =
-        getBlockedRelayListFlow()
-            .map {
-                normalizeBlockedRelayListWithBackup(it.note)
-            }.onStart { emit(normalizeBlockedRelayListWithBackup(blockedListNote)) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-
-    suspend fun saveRelayList(blockedRelays: List<NormalizedRelayUrl>): BlockedRelayListEvent {
-        if (!signer.isWriteable()) throw SignerExceptions.ReadOnlyException()
-        val relayListForBlocked = getBlockedRelayList()
-
-        return if (relayListForBlocked != null && relayListForBlocked.tags.isNotEmpty()) {
-            BlockedRelayListEvent.updateRelayList(
-                earlierVersion = relayListForBlocked,
-                relays = blockedRelays,
-                signer = signer,
-            )
-        } else {
-            BlockedRelayListEvent.create(
-                relays = blockedRelays,
-                signer = signer,
-            )
-        }
-    }
-
-    init {
-        settings.backupBlockedRelayList?.let {
-            Log.d("AccountRegisterObservers") { "Loading saved Blocked relay list ${it.toJson()}" }
-            @OptIn(DelicateCoroutinesApi::class)
-            scope.launch(Dispatchers.IO) { LocalCache.justConsumeMyOwnEvent(it) }
-        }
-
-        scope.launch(Dispatchers.IO) {
-            Log.d("AccountRegisterObservers", "Blocked Relay List Collector Start")
-            getBlockedRelayListFlow().collect {
-                Log.d("AccountRegisterObservers") { "Updating Blocked Relay List for ${signer.pubKey}" }
-                (it.note.event as? BlockedRelayListEvent)?.let {
-                    settings.updateBlockedRelayList(it)
-                }
-            }
-        }
-    }
-}
+typealias BlockedRelayListState = com.vitorpamplona.amethyst.commons.model.nip51Lists.blockedRelays.BlockedRelayListState
