+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/broadcastRelays/BroadcastRelayListState.kt
@@ -20,68 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.broadcastRelays
 
-import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.NoteState
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.relayLists.BroadcastRelayListEvent
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-
-class BroadcastRelayListState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val decryptionCache: BroadcastRelayListDecryptionCache,
-    val scope: CoroutineScope,
-    val settings: AccountSettings,
-) {
-    // Creates a long-term reference for this note so that the GC doesn't collect the note it self
-    val broadcastListNote = cache.getOrCreateAddressableNote(getBroadcastRelayListAddress())
-
-    fun getBroadcastRelayListAddress() = BroadcastRelayListEvent.createAddress(signer.pubKey)
-
-    fun getBroadcastRelayListFlow(): StateFlow<NoteState> = broadcastListNote.flow().metadata.stateFlow
-
-    fun getBroadcastRelayList(): BroadcastRelayListEvent? = broadcastListNote.event as? BroadcastRelayListEvent
-
-    suspend fun normalizeBroadcastRelayListWithBackup(note: Note): Set<NormalizedRelayUrl> {
-        val event = note.event as? BroadcastRelayListEvent
-        return event?.let { decryptionCache.relays(it) } ?: emptySet()
-    }
-
-    val flow =
-        getBroadcastRelayListFlow()
-            .map { normalizeBroadcastRelayListWithBackup(it.note) }
-            .onStart { emit(normalizeBroadcastRelayListWithBackup(broadcastListNote)) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-
-    suspend fun saveRelayList(broadcastRelays: List<NormalizedRelayUrl>): BroadcastRelayListEvent {
-        val relayListForBroadcast = getBroadcastRelayList()
-
-        return if (relayListForBroadcast != null && relayListForBroadcast.tags.isNotEmpty()) {
-            BroadcastRelayListEvent.updateRelayList(
-                earlierVersion = relayListForBroadcast,
-                relays = broadcastRelays,
-                signer = signer,
-            )
-        } else {
-            BroadcastRelayListEvent.create(
-                relays = broadcastRelays,
-                signer = signer,
-            )
-        }
-    }
-}
+typealias BroadcastRelayListState = com.vitorpamplona.amethyst.commons.model.nip51Lists.broadcastRelays.BroadcastRelayListState
