+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/geohashLists/GeohashListState.kt
@@ -20,107 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.geohashLists
 
-import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.NoteState
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.geohashList.GeohashListEvent
-import com.vitorpamplona.quartz.utils.Log
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.flow.transformLatest
-import kotlinx.coroutines.launch
-
-class GeohashListState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val decryptionCache: GeohashListDecryptionCache,
-    val scope: CoroutineScope,
-    val settings: AccountSettings,
-) {
-    // Creates a long-term reference for this note so that the GC doesn't collect the note it self
-    val geohashListNote = cache.getOrCreateAddressableNote(getGeohashListAddress())
-
-    fun getGeohashListAddress() = GeohashListEvent.createAddress(signer.pubKey)
-
-    fun getGeohashListFlow(): StateFlow<NoteState> = geohashListNote.flow().metadata.stateFlow
-
-    fun getGeohashList(): GeohashListEvent? = geohashListNote.event as? GeohashListEvent
-
-    suspend fun geohashListWithBackup(note: Note): Set<String> {
-        val event = note.event as? GeohashListEvent ?: settings.backupGeohashList
-        return event?.let { decryptionCache.geohashes(it).mapTo(mutableSetOf()) { it.lowercase() } } ?: emptySet()
-    }
-
-    @OptIn(ExperimentalCoroutinesApi::class)
-    val flow: StateFlow<Set<String>> =
-        getGeohashListFlow()
-            .transformLatest { noteState ->
-                emit(geohashListWithBackup(noteState.note))
-            }.onStart {
-                emit(geohashListWithBackup(geohashListNote))
-            }.flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-
-    suspend fun follow(geohashes: List<String>): GeohashListEvent {
-        val geohashList = getGeohashList()
-
-        return if (geohashList == null) {
-            GeohashListEvent.create(geohashes, true, signer)
-        } else {
-            GeohashListEvent.add(geohashList, geohashes, true, signer)
-        }
-    }
-
-    suspend fun follow(geohash: String): GeohashListEvent {
-        val geohashList = getGeohashList()
-
-        return if (geohashList == null) {
-            GeohashListEvent.create(geohash, true, signer)
-        } else {
-            GeohashListEvent.add(geohashList, geohash, true, signer)
-        }
-    }
-
-    suspend fun unfollow(geohash: String): GeohashListEvent? {
-        val geohashList = getGeohashList()
-
-        return if (geohashList != null) {
-            GeohashListEvent.remove(geohashList, geohash, signer)
-        } else {
-            null
-        }
-    }
-
-    init {
-        settings.backupGeohashList?.let { event ->
-            Log.d("AccountRegisterObservers") { "Loading saved Geohash list ${event.toJson()}" }
-            @OptIn(DelicateCoroutinesApi::class)
-            scope.launch(Dispatchers.IO) {
-                LocalCache.justConsumeMyOwnEvent(event)
-            }
-        }
-
-        scope.launch(Dispatchers.IO) {
-            Log.d("AccountRegisterObservers", "Geohash List Collector Start")
-            getGeohashListFlow().collect { noteState ->
-                Log.d("AccountRegisterObservers") { "Geohash List for ${signer.pubKey}" }
-                (noteState.note.event as? GeohashListEvent)?.let {
-                    settings.updateGeohashListTo(it)
-                }
-            }
-        }
-    }
-}
+typealias GeohashListState = com.vitorpamplona.amethyst.commons.model.nip51Lists.geohashLists.GeohashListState
