+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/muteList/MuteListState.kt
@@ -20,142 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.muteList
 
-import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.NoteState
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.muteList.MuteListEvent
-import com.vitorpamplona.quartz.nip51Lists.muteList.tags.MuteTag
-import com.vitorpamplona.quartz.nip51Lists.muteList.tags.UserTag
-import com.vitorpamplona.quartz.nip51Lists.muteList.tags.WordTag
-import com.vitorpamplona.quartz.utils.Log
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.launch
-
-class MuteListState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val decryptionCache: MuteListDecryptionCache,
-    val scope: CoroutineScope,
-    val settings: AccountSettings,
-) {
-    // Creates a long-term reference for this note so that the GC doesn't collect the note it self
-    val muteListNote = cache.getOrCreateAddressableNote(getMuteListAddress())
-
-    fun getMuteListAddress() = MuteListEvent.createAddress(signer.pubKey)
-
-    fun getMuteListFlow(): StateFlow<NoteState> = muteListNote.flow().metadata.stateFlow
-
-    fun getMuteList(): MuteListEvent? = muteListNote.event as? MuteListEvent
-
-    suspend fun muteListWithBackup(note: Note): List<MuteTag> {
-        val event = note.event as? MuteListEvent ?: settings.backupMuteList
-        return event?.let { decryptionCache.mutedUsersAndWords(it) } ?: emptyList()
-    }
-
-    val flow =
-        getMuteListFlow()
-            .map { muteListWithBackup(it.note) }
-            .onStart { emit(muteListWithBackup(muteListNote)) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptyList(),
-            )
-
-    suspend fun hideUser(pubkeyHex: String): MuteListEvent {
-        val muteList = getMuteList()
-
-        return if (muteList != null) {
-            MuteListEvent.add(
-                earlierVersion = muteList,
-                mute = UserTag(pubkeyHex),
-                isPrivate = true,
-                signer = signer,
-            )
-        } else {
-            MuteListEvent.create(
-                mute = UserTag(pubkeyHex),
-                isPrivate = true,
-                signer = signer,
-            )
-        }
-    }
-
-    suspend fun showUser(pubkeyHex: String): MuteListEvent? {
-        val muteList = getMuteList()
-
-        return if (muteList != null) {
-            MuteListEvent.remove(
-                earlierVersion = muteList,
-                mute = UserTag(pubkeyHex),
-                signer = signer,
-            )
-        } else {
-            null
-        }
-    }
-
-    suspend fun hideWord(word: String): MuteListEvent {
-        val muteList = getMuteList()
-
-        return if (muteList != null) {
-            MuteListEvent.add(
-                earlierVersion = muteList,
-                mute = WordTag(word),
-                isPrivate = true,
-                signer = signer,
-            )
-        } else {
-            MuteListEvent.create(
-                mute = WordTag(word),
-                isPrivate = true,
-                signer = signer,
-            )
-        }
-    }
-
-    suspend fun showWord(word: String): MuteListEvent? {
-        val muteList = getMuteList()
-
-        return if (muteList != null) {
-            MuteListEvent.remove(
-                earlierVersion = muteList,
-                mute = WordTag(word),
-                signer = signer,
-            )
-        } else {
-            null
-        }
-    }
-
-    init {
-        settings.backupMuteList?.let { event ->
-            Log.d("AccountRegisterObservers") { "Loading saved mute list ${event.toJson()}" }
-            @OptIn(DelicateCoroutinesApi::class)
-            scope.launch(Dispatchers.IO) {
-                LocalCache.justConsumeMyOwnEvent(event)
-            }
-        }
-
-        scope.launch(Dispatchers.IO) {
-            Log.d("AccountRegisterObservers", "Mute List Collector Start")
-            getMuteListFlow().collect {
-                Log.d("AccountRegisterObservers") { "Updating Mute List for ${signer.pubKey}" }
-                (it.note.event as? MuteListEvent)?.let {
-                    settings.updateMuteList(it)
-                }
-            }
-        }
-    }
-}
+typealias MuteListState = com.vitorpamplona.amethyst.commons.model.nip51Lists.muteList.MuteListState
