+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/proxyRelays/ProxyRelayListState.kt
@@ -20,68 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.proxyRelays
 
-import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.NoteState
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.relayLists.ProxyRelayListEvent
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-
-class ProxyRelayListState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val decryptionCache: ProxyRelayListDecryptionCache,
-    val scope: CoroutineScope,
-    val settings: AccountSettings,
-) {
-    // Creates a long-term reference for this note so that the GC doesn't collect the note it self
-    val proxyListNote = cache.getOrCreateAddressableNote(getProxyRelayListAddress())
-
-    fun getProxyRelayListAddress() = ProxyRelayListEvent.createAddress(signer.pubKey)
-
-    fun getProxyRelayListFlow(): StateFlow<NoteState> = proxyListNote.flow().metadata.stateFlow
-
-    fun getProxyRelayList(): ProxyRelayListEvent? = proxyListNote.event as? ProxyRelayListEvent
-
-    suspend fun normalizeProxyRelayListWithBackup(note: Note): Set<NormalizedRelayUrl> {
-        val event = note.event as? ProxyRelayListEvent
-        return event?.let { decryptionCache.relays(it) } ?: emptySet()
-    }
-
-    val flow =
-        getProxyRelayListFlow()
-            .map { normalizeProxyRelayListWithBackup(it.note) }
-            .onStart { emit(normalizeProxyRelayListWithBackup(proxyListNote)) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-
-    suspend fun saveRelayList(proxyRelays: List<NormalizedRelayUrl>): ProxyRelayListEvent {
-        val relayListForProxy = getProxyRelayList()
-
-        return if (relayListForProxy != null && relayListForProxy.tags.isNotEmpty()) {
-            ProxyRelayListEvent.updateRelayList(
-                earlierVersion = relayListForProxy,
-                relays = proxyRelays,
-                signer = signer,
-            )
-        } else {
-            ProxyRelayListEvent.create(
-                relays = proxyRelays,
-                signer = signer,
-            )
-        }
-    }
-}
+typealias ProxyRelayListState = com.vitorpamplona.amethyst.commons.model.nip51Lists.proxyRelays.ProxyRelayListState
