+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/relayFeeds/RelayFeedListState.kt
@@ -20,114 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.relayFeeds
 
-import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.NoteState
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.relayLists.RelayFeedsListEvent
-import com.vitorpamplona.quartz.utils.Log
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.launch
-
-class RelayFeedListState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val decryptionCache: RelayFeedsListDecryptionCache,
-    val scope: CoroutineScope,
-    val settings: AccountSettings,
-) {
-    // Creates a long-term reference for this note so that the GC doesn't collect the note it self
-    val relayFeedListNote = cache.getOrCreateAddressableNote(getRelayFeedsListAddress())
-
-    fun getRelayFeedsListAddress() = RelayFeedsListEvent.createAddress(signer.pubKey)
-
-    fun getRelayFeedsListFlow(): StateFlow<NoteState> = relayFeedListNote.flow().metadata.stateFlow
-
-    fun getRelayFeedsList(): RelayFeedsListEvent? = relayFeedListNote.event as? RelayFeedsListEvent
-
-    fun relayFeedsListEvent(note: Note) = note.event as? RelayFeedsListEvent ?: settings.backupRelayFeedsList
-
-    suspend fun normalizeRelayFeedsListWithBackup(note: Note): Set<NormalizedRelayUrl> = relayFeedsListEvent(note)?.let { decryptionCache.relays(it) }?.ifEmpty { null } ?: emptySet()
-
-    suspend fun normalizeRelayFeedsListWithBackupNoDefaults(note: Note): Set<NormalizedRelayUrl> = relayFeedsListEvent(note)?.let { decryptionCache.relays(it) } ?: emptySet()
-
-    val flow =
-        getRelayFeedsListFlow()
-            .map { normalizeRelayFeedsListWithBackup(it.note) }
-            .onStart { emit(normalizeRelayFeedsListWithBackup(relayFeedListNote)) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-
-    val flowNoDefaults =
-        getRelayFeedsListFlow()
-            .map { normalizeRelayFeedsListWithBackupNoDefaults(it.note) }
-            .onStart { emit(normalizeRelayFeedsListWithBackupNoDefaults(relayFeedListNote)) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-
-    suspend fun addRelay(relay: NormalizedRelayUrl): RelayFeedsListEvent {
-        val current = normalizeRelayFeedsListWithBackupNoDefaults(relayFeedListNote).toMutableList()
-        if (relay !in current) current.add(relay)
-        return saveRelayList(current)
-    }
-
-    suspend fun removeRelay(relay: NormalizedRelayUrl): RelayFeedsListEvent? {
-        val current = normalizeRelayFeedsListWithBackupNoDefaults(relayFeedListNote).toMutableList()
-        if (relay !in current) return null
-        current.remove(relay)
-        return saveRelayList(current)
-    }
-
-    suspend fun saveRelayList(relayFeeds: List<NormalizedRelayUrl>): RelayFeedsListEvent {
-        val relayFeedsList = getRelayFeedsList()
-
-        return if (relayFeedsList != null && relayFeedsList.tags.isNotEmpty()) {
-            RelayFeedsListEvent.updateRelayList(
-                earlierVersion = relayFeedsList,
-                relays = relayFeeds,
-                signer = signer,
-            )
-        } else {
-            RelayFeedsListEvent.create(
-                relays = relayFeeds,
-                signer = signer,
-            )
-        }
-    }
-
-    init {
-        settings.backupRelayFeedsList?.let {
-            Log.d("AccountRegisterObservers") { "Loading saved relay feeds list ${it.toJson()}" }
-            @OptIn(DelicateCoroutinesApi::class)
-            scope.launch(Dispatchers.IO) { LocalCache.justConsumeMyOwnEvent(it) }
-        }
-
-        scope.launch(Dispatchers.IO) {
-            Log.d("AccountRegisterObservers", "Relay feeds list Collector Start")
-            getRelayFeedsListFlow().collect {
-                Log.d("AccountRegisterObservers") { "Updating Relay feeds list for ${signer.pubKey}" }
-                (it.note.event as? RelayFeedsListEvent)?.let {
-                    settings.updateRelayFeedList(it)
-                }
-            }
-        }
-    }
-}
+typealias RelayFeedListState = com.vitorpamplona.amethyst.commons.model.nip51Lists.relayFeeds.RelayFeedListState
