+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip51Lists/trustedRelays/TrustedRelayListState.kt
@@ -20,89 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip51Lists.trustedRelays
 
-import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.Note
-import com.vitorpamplona.amethyst.model.NoteState
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip51Lists.relayLists.TrustedRelayListEvent
-import com.vitorpamplona.quartz.utils.Log
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.flowOn
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onStart
-import kotlinx.coroutines.flow.stateIn
-import kotlinx.coroutines.launch
-
-class TrustedRelayListState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val decryptionCache: TrustedRelayListDecryptionCache,
-    val scope: CoroutineScope,
-    val settings: AccountSettings,
-) {
-    // Creates a long-term reference for this note so that the GC doesn't collect the note it self
-    val trustedListNote = cache.getOrCreateAddressableNote(getTrustedRelayListAddress())
-
-    fun getTrustedRelayListAddress() = TrustedRelayListEvent.createAddress(signer.pubKey)
-
-    fun getTrustedRelayListFlow(): StateFlow<NoteState> = trustedListNote.flow().metadata.stateFlow
-
-    fun getTrustedRelayList(): TrustedRelayListEvent? = trustedListNote.event as? TrustedRelayListEvent
-
-    suspend fun normalizeTrustedRelayListWithBackup(note: Note): Set<NormalizedRelayUrl> {
-        val event = note.event as? TrustedRelayListEvent ?: settings.backupTrustedRelayList
-        return event?.let { decryptionCache.relays(it) } ?: emptySet()
-    }
-
-    val flow =
-        getTrustedRelayListFlow()
-            .map { normalizeTrustedRelayListWithBackup(it.note) }
-            .onStart { emit(normalizeTrustedRelayListWithBackup(trustedListNote)) }
-            .flowOn(Dispatchers.IO)
-            .stateIn(
-                scope,
-                SharingStarted.Eagerly,
-                emptySet(),
-            )
-
-    suspend fun saveRelayList(trustedRelays: List<NormalizedRelayUrl>): TrustedRelayListEvent {
-        val relayListForTrusted = getTrustedRelayList()
-
-        return if (relayListForTrusted != null && relayListForTrusted.tags.isNotEmpty()) {
-            TrustedRelayListEvent.updateRelayList(
-                earlierVersion = relayListForTrusted,
-                relays = trustedRelays,
-                signer = signer,
-            )
-        } else {
-            TrustedRelayListEvent.create(
-                relays = trustedRelays,
-                signer = signer,
-            )
-        }
-    }
-
-    init {
-        settings.backupTrustedRelayList?.let {
-            Log.d("AccountRegisterObservers") { "Loading saved Trusted relay list ${it.toJson()}" }
-            @OptIn(DelicateCoroutinesApi::class)
-            scope.launch(Dispatchers.IO) { cache.justConsumeMyOwnEvent(it) }
-        }
-
-        scope.launch(Dispatchers.IO) {
-            Log.d("AccountRegisterObservers", "Trusted Relay List Collector Start")
-            getTrustedRelayListFlow().collect {
-                Log.d("AccountRegisterObservers") { "Updating Trusted Relay List for ${signer.pubKey}" }
-                (it.note.event as? TrustedRelayListEvent)?.let {
-                    settings.updateTrustedRelayList(it)
-                }
-            }
-        }
-    }
-}
+typealias TrustedRelayListState = com.vitorpamplona.amethyst.commons.model.nip51Lists.trustedRelays.TrustedRelayListState
