+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/nip78AppSpecific/AppSpecificState.kt
@@ -20,87 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.nip78AppSpecific
 
-import com.vitorpamplona.amethyst.model.AccountSettings
-import com.vitorpamplona.amethyst.model.AccountSyncedSettingsInternal
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.NoteState
-import com.vitorpamplona.quartz.nip01Core.core.JsonMapper
-import com.vitorpamplona.quartz.nip01Core.signers.NostrSigner
-import com.vitorpamplona.quartz.nip78AppData.AppSpecificDataEvent
-import com.vitorpamplona.quartz.utils.Log
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.DelicateCoroutinesApi
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
-import kotlin.coroutines.cancellation.CancellationException
-
-class AppSpecificState(
-    val signer: NostrSigner,
-    val cache: LocalCache,
-    val scope: CoroutineScope,
-    val settings: AccountSettings,
-) {
-    companion object {
-        const val APP_SPECIFIC_DATA_D_TAG = "AmethystSettings"
-    }
-
-    // Creates a long-term reference for this note so that the GC doesn't collect the note it self
-    val amethystSettingsNote = cache.getOrCreateAddressableNote(getAppSpecificDataAddress())
-
-    fun getAppSpecificDataAddress() = AppSpecificDataEvent.createAddress(signer.pubKey, APP_SPECIFIC_DATA_D_TAG)
-
-    fun getAppSpecificDataFlow(): StateFlow<NoteState> = amethystSettingsNote.flow().metadata.stateFlow
-
-    suspend fun saveNewAppSpecificData(): AppSpecificDataEvent {
-        val toInternal = settings.syncedSettings.toInternal()
-        return AppSpecificDataEvent.create(
-            dTag = APP_SPECIFIC_DATA_D_TAG,
-            description = signer.nip44Encrypt(JsonMapper.toJson(toInternal), signer.pubKey),
-            otherTags = emptyArray(),
-            signer = signer,
-        )
-    }
-
-    init {
-        if (settings.isWriteable()) {
-            settings.backupAppSpecificData?.let { event ->
-                Log.d("AccountRegisterObservers") { "Loading saved app specific data ${event.toJson()}" }
-                @OptIn(DelicateCoroutinesApi::class)
-                scope.launch(Dispatchers.IO) {
-                    LocalCache.justConsumeMyOwnEvent(event)
-                    try {
-                        val decrypted = signer.decrypt(event.content, event.pubKey)
-                        val syncedSettings = JsonMapper.fromJson<AccountSyncedSettingsInternal>(decrypted)
-                        settings.syncedSettings.updateFrom(syncedSettings)
-                    } catch (e: Throwable) {
-                        if (e is CancellationException) throw e
-                        Log.w("LocalPreferences", "Error Decoding latestAppSpecificData from Preferences with value", e)
-                    }
-                }
-            }
-
-            scope.launch(Dispatchers.IO) {
-                Log.d("AccountRegisterObservers", "AppSpecificData Collector Start")
-                getAppSpecificDataFlow().collect {
-                    try {
-                        Log.d("AccountRegisterObservers") { "Updating AppSpecificData for ${signer.pubKey}" }
-                        (it.note.event as? AppSpecificDataEvent)?.let {
-                            val decrypted = signer.decrypt(it.content, it.pubKey)
-                            try {
-                                val syncedSettings = JsonMapper.fromJson<AccountSyncedSettingsInternal>(decrypted)
-                                settings.updateAppSpecificData(it, syncedSettings)
-                            } catch (e: Throwable) {
-                                if (e is CancellationException) throw e
-                                Log.w("LocalPreferences", "Error Decoding latestAppSpecificData from Preferences with value $decrypted", e)
-                            }
-                        }
-                    } catch (e: Throwable) {
-                        if (e is CancellationException) throw e
-                        Log.w("LocalPreferences", "Error Decrypting latestAppSpecificData from Preferences", e)
-                    }
-                }
-            }
-        }
-    }
-}
+typealias AppSpecificState = com.vitorpamplona.amethyst.commons.model.nip78AppSpecific.AppSpecificState
