+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/model/localRelays/ForwardKind0ToLocalRelayState.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2025 Vitor Pamplona
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+ * Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
+ * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package com.vitorpamplona.amethyst.commons.model.localRelays
+
+import com.vitorpamplona.amethyst.commons.model.AccountSettings
+import com.vitorpamplona.quartz.nip01Core.metadata.MetadataEvent
+import com.vitorpamplona.quartz.nip01Core.relay.client.INostrClient
+import com.vitorpamplona.quartz.nip01Core.relay.client.accessories.EventCollector
+
+class ForwardKind0ToLocalRelayState(
+    val client: INostrClient,
+    val localRelayList: LocalRelayListState,
+    val settings: AccountSettings,
+) {
+    private val eventCollector =
+        EventCollector(client) { event, _ ->
+            if (event is MetadataEvent && settings.syncedSettings.security.sendKind0EventsToLocalRelay.value) {
+                val localRelays = localRelayList.flow.value
+                if (localRelays.isNotEmpty()) {
+                    client.publish(event, localRelays)
+                }
+            }
+        }
+}
